Change all yaml keys to strings.
Symbols in yaml are ugly. This commit changes symbols in the yaml files to strings and adds a function to the Vagrantfile to convert the keys to symbols for those instances where they are required by Varant's syntax. I also did a little formatting cleanup while I was at it.