Switch provider setting sections.
The general provider settings should never change, while the provider-specific settings could eventually grow into a large block of settings for the various providers. I put the provider-specific portion at the end of the Vagrantfile so it won't clutter things up too much if that happens (though I'll probably just pull it out in that case).